+++ policyengine_uk/parameters/gov/dwp/pension_credit/savings_credit/cutoff_year.yaml
@@ -1,0 +1,9 @@
+description: The year after which new pensioners are not eligible for Savings Credit.
+metadata:
+  name: pc_sc_cutoff_year
+  label: Savings Credit eligibility cutoff year
+  reference:
+    - title: State Pension Credit Act 2002
+      href: https://www.legislation.gov.uk/ukpga/2002/16/section/3
+values:
+  2002-01-01: 2016
