+++ policyengine_uk/variables/gov/dwp/pension_credit/savings_credit/is_savings_credit_eligible.py
@@ -10,13 +10,13 @@ class is_savings_credit_eligible(Variable):
     reference = "https://www.legislation.gov.uk/ukpga/2002/16/section/3"
 
     def formula(benunit, period, parameters):
-        # Check if any member reached SPA before 2016
-        has_pre_2016_spa_member = benunit.any(
-            benunit.members("reached_SPA_before_2016", period)
+        # Check if any member reached SPA before the cutoff year
+        has_pre_cutoff_spa_member = benunit.any(
+            benunit.members("meets_savings_credit_age_requirement", period)
         )
         income = benunit("savings_credit_income", period)
         sc = parameters(period).gov.dwp.pension_credit.savings_credit
         relation_type = benunit("relation_type", period)
         threshold = sc.threshold[relation_type]
         meets_income_threshold = income >= threshold
-        return has_pre_2016_spa_member & meets_income_threshold
+        return has_pre_cutoff_spa_member & meets_income_threshold
